% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invalid_utf8.R
\name{check_invalid_utf8}
\alias{check_invalid_utf8}
\title{Check Invalid UTF-8}
\usage{
check_invalid_utf8(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a logical vector
}
\description{
Check Invalid UTF-8
}
\examples{
x <- c("Elm Grove\xa0, 505 Southampton Street East, Hastings", "123")
check_invalid_utf8(x)
}
