% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invalid_utf8.R
\name{convert_to_utf8}
\alias{convert_to_utf8}
\title{Convert to UTF-8}
\usage{
convert_to_utf8(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector
}
\description{
Convert to UTF-8
}
\examples{
x <- c("Elm Grove\xa0, 505 Southampton Street East, Hastings", "123")
convert_to_utf8(x)
}
