% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_df_key_info.R
\name{db_tbl_key_info}
\alias{db_tbl_key_info}
\title{Find Key Info of a Table in Database}
\usage{
db_tbl_key_info(db_con, tbl_name)
}
\arguments{
\item{db_con}{database connection}

\item{tbl_name}{table's name}
}
\value{
a dataframe
}
\description{
Find Key Info of a Table in Database
}
