% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_df_key_info.R
\name{db_tbl_names}
\alias{db_tbl_names}
\title{Find Names of Tables in Database}
\usage{
db_tbl_names(db_con)
}
\arguments{
\item{db_con}{database connection}
}
\value{
a vector
}
\description{
Find Names of Tables in Database
}
