% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_df_key_info.R
\name{find_df_key_info}
\alias{find_df_key_info}
\title{Find Data Frame Key Information}
\usage{
find_df_key_info(a_df)
}
\arguments{
\item{a_df}{a data frame}
}
\value{
a data frame
}
\description{
This is to find key information of a data frame for understanding/checking
the data frame
}
\examples{
temp_df <-
     data.frame(a = 1:10,
                b = NA,
                e = c(letters[1:8], NA, NA),
                stringsAsFactors = FALSE)
x <- find_df_key_info(temp_df)
}
