% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_sums_added.R
\name{get_col_sums_added}
\alias{get_col_sums_added}
\title{Get Column Sums Added In Dataframe}
\usage{
get_col_sums_added(a_df, first_col_is_label = TRUE)
}
\arguments{
\item{a_df}{a data frame}

\item{first_col_is_label}{logical variable, default value TRUE}
}
\value{
a data frame, where the last row has column sums if the column is
summable
}
\description{
This is to get column sums added in dataframe as its last row
}
\examples{
my_df <- data.frame(id = letters[1:5],
x = c(1:3, 4.1, NA),
y = LETTERS[1:5],
z = c(NA, NA, 1:3),
w = rep(NA, 5),
u = rep("", 5)
)
new_df <- get_col_sums_added(a_df = my_df)
}
