% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{modify}
\alias{modify}
\alias{modi_sum}
\alias{modi_min}
\alias{modi_max}
\alias{min_length}
\alias{max_length}
\title{Modified Sum, Min, Max functions (written by Matthew Hendtlass)}
\usage{
modify(f, default = NA_real_)

modi_sum(x)

modi_min(x)

modi_max(x)

min_length(x)

max_length(x)
}
\arguments{
\item{f}{an R function}

\item{default}{default value}

\item{x}{a vector}
}
\value{
A numeric vector of length 1
}
\description{
Applies operation with NAs removed; if the input is either not numeric or all
\code{NAs}, \code{modi_sum, modi_min, modi_max} return \code{NA_real_} and
\code{min_length, max_length} return \code{0}.
}
