% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_spaces_in_df.R
\name{my_trim_ws}
\alias{my_trim_ws}
\title{My Trim White Space}
\usage{
my_trim_ws(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector---the input itself or after trimming
}
\description{
Get rid of spaces: no sapce at the begining and ending; leave only one space
between characters
}
\examples{
x <- ' long   space  character  '
y <- my_trim_ws(x)
}
