% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_spaces_in_df.R
\name{trim_spaces_in_df}
\alias{trim_spaces_in_df}
\title{Trim Spaces in Data Frame}
\usage{
trim_spaces_in_df(a_df)
}
\arguments{
\item{a_df}{a data frame}
}
\value{
a data frame
}
\description{
Trim Spaces in Data Frame
}
\examples{
df <- data.frame(x = 1:4,
                 y = c('a b ', ' cd e ', 'fgh i ', 'ok'),
                 z = rep(TRUE, 4))
new_df <- trim_spaces_in_df(df)
}
