% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_db.R
\name{write_a_file_to_duckdb}
\alias{write_a_file_to_duckdb}
\title{Write a File to Duckdb}
\usage{
write_a_file_to_duckdb(data, the_dir, table_name, overwrite_indi = FALSE)
}
\arguments{
\item{data}{the data, normally a dataframe}

\item{the_dir}{the directory where the database locates}

\item{table_name}{name of table to be created}

\item{overwrite_indi}{logical, default is FALSE}
}
\description{
Write a File to Duckdb
}
